+++ CHANGELOG.md
@@ -13,11 +13,13 @@ All notable changes to this project will be documented in this file[^1].
 ### Fixed
 - Username not fillable in admin
 - Do not overwrite authority links from CEDVU
+- Harvester error logging
+- Harvester missing total in progress
 
 ## [2.27.0] - 2021-04-16
 ### Fixed
  - Counting clicks on carousel slides
- 
+
 ### Changed
 - Upgrade to Laravel 7
 - Upgrade to Laravel 8